+++ test/LionWeb.Core.Test/NodeApi/Lenient/Containment/Multiple/Required/NodeVariantsTests.cs
@@ -71,7 +71,6 @@ public class NodeVariantsTests : LenientNodeTestsBase
     }
 
     [TestMethod]
-    [Ignore("fails after adding generic api")]
     public void IReadableNode_Reflective()
     {
         var parent = newCompositeShape("g");